+++ cookbooks/cwb-server/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.7.4 (2020-04-21)
+
+* Add workaround for broken `vagrant-aws` installation (see [vagrant issue](https://github.com/hashicorp/vagrant/issues/11518))
+* Fix unpinned bundler version
+* Update cookbook dependencies
+
 # 0.7.3 (2019-01-03)
 
 * Expose `SSH_PUB_KEY_PATH`